+++ internal/nodetree/nodetree.go
@@ -41,6 +41,7 @@ type (
 
 		EndNS       uint64                                `json:"-"`
 		Frame       frame.Frame                           `json:"-"`
+		Occurrence  int                                   `json:"-"`
 		SampleCount int                                   `json:"-"`
 		StartNS     uint64                                `json:"-"`
 		ProfileIDs  map[string]struct{}                   `json:"profile_ids,omitempty"`
@@ -62,6 +63,7 @@ func NodeFromFrame(f frame.Frame, start, end, fingerprint uint64) *Node {
 		Name:          f.Function,
 		Package:       f.ModuleOrPackage(),
 		Path:          f.Path,
+		Occurrence:    1,
 		SampleCount:   1,
 		StartNS:       start,
 		ProfileIDs:    map[string]struct{}{},